+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/LazyVersion.kt
@@ -1,0 +1,8 @@
+package com.javiersc.semver.gradle.plugin
+
+import org.gradle.api.provider.Provider
+
+public class LazyVersion(internal val version: Provider<String>) {
+
+    override fun toString(): String = version.get()
+}