+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/internal/AppliedOnlyOnRootProject.kt
@@ -1,6 +1,0 @@
-package com.javiersc.semver.gradle.plugin.internal
-
-import org.gradle.api.Project
-
-internal val Project.appliedOnlyOnRootProject: Boolean
-    get() = rootProject.hasSemVerPlugin && rootProject.subprojects.none(Project::hasSemVerPlugin)