+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/internal/CalculatedVersion.kt
@@ -2,28 +2,28 @@ package com.javiersc.semver.gradle.plugin.internal
 
 import com.javiersc.semver.Version
 import com.javiersc.semver.Version.Increase
-import java.util.Date
+import com.javiersc.semver.gradle.plugin.internal.git.GitRef
+import com.javiersc.semver.gradle.plugin.internal.git.commitsBetweenTwoCommitsIncludingLastExcludingFirst
+import com.javiersc.semver.gradle.plugin.internal.git.headCommit
+import com.javiersc.semver.gradle.plugin.internal.git.isThereVersionTag
+import com.javiersc.semver.gradle.plugin.internal.git.lastCommitInCurrentBranch
+import com.javiersc.semver.gradle.plugin.internal.git.lastVersionCommitInCurrentBranch
+import com.javiersc.semver.gradle.plugin.internal.git.lastVersionInCurrentBranch
 import org.eclipse.jgit.api.Git
 
 @Suppress("ComplexMethod")
 internal fun Git.calculatedVersion(
-    warningLastVersionIsNotHigherVersion: (last: Version?, higher: Version?) -> Unit,
     tagPrefix: String,
     stageProperty: String?,
     scopeProperty: String?,
     isCreatingSemverTag: Boolean,
-    mockDate: Date?,
     checkClean: Boolean,
 ): String {
-    val lastSemVer =
-        lastVersionInCurrentBranch(
-            warningLastVersionIsNotHigherVersion = warningLastVersionIsNotHigherVersion,
-            tagPrefix = tagPrefix,
-        )
+    val lastSemver: Version = lastVersionInCurrentBranch(tagPrefix = tagPrefix)
 
-    val previousStage: String? = lastSemVer.stage?.name
+    val previousStage: String? = lastSemver.stage?.name
 
-    val incStage =
+    val incStage: String =
         (stageProperty ?: previousStage ?: "").run {
             when {
                 equals(Stage.Auto(), true) && !previousStage.isNullOrBlank() -> previousStage
@@ -31,41 +31,45 @@ internal fun Git.calculatedVersion(
                 else -> this
             }
         }
+
+    val isNoStagedNoScopedNoCreatingSemverTag: Boolean =
+        stageProperty.isNullOrBlank() && scopeProperty.isNullOrBlank() && !isCreatingSemverTag
+
+    val isCleanAndShouldBeChecked: Boolean = status().call().isClean.not() && checkClean
+
     val calculatedVersion: String =
         when {
-            (stageProperty.isNullOrBlank() &&
-                scopeProperty.isNullOrBlank() &&
-                !isCreatingSemverTag) || status().call().isClean.not() && checkClean -> {
-                "$lastSemVer${calculateAdditionalVersionData(tagPrefix, mockDate, checkClean)}"
+            isNoStagedNoScopedNoCreatingSemverTag || isCleanAndShouldBeChecked -> {
+                "$lastSemver${calculateAdditionalVersionData(tagPrefix, checkClean)}"
             }
             stageProperty.equals(Stage.Snapshot(), ignoreCase = true) -> {
                 when (scopeProperty) {
-                    Scope.Major() -> "${lastSemVer.nextSnapshotMajor()}"
-                    Scope.Minor() -> "${lastSemVer.nextSnapshotMinor()}"
-                    Scope.Patch() -> "${lastSemVer.nextSnapshotPatch()}"
-                    else -> "${lastSemVer.nextSnapshotPatch()}"
+                    Scope.Major() -> "${lastSemver.nextSnapshotMajor()}"
+                    Scope.Minor() -> "${lastSemver.nextSnapshotMinor()}"
+                    Scope.Patch() -> "${lastSemver.nextSnapshotPatch()}"
+                    else -> "${lastSemver.nextSnapshotPatch()}"
                 }
             }
             stageProperty.equals(Stage.Final(), ignoreCase = true) -> {
                 when (scopeProperty) {
-                    Scope.Major() -> "${lastSemVer.inc(Increase.Major, "")}"
-                    Scope.Minor() -> "${lastSemVer.inc(Increase.Minor, "")}"
-                    Scope.Patch() -> "${lastSemVer.inc(Increase.Patch, "")}"
-                    else -> "${lastSemVer.inc(stageName = "")}"
+                    Scope.Major() -> "${lastSemver.inc(Increase.Major, "")}"
+                    Scope.Minor() -> "${lastSemver.inc(Increase.Minor, "")}"
+                    Scope.Patch() -> "${lastSemver.inc(Increase.Patch, "")}"
+                    else -> "${lastSemver.inc(stageName = "")}"
                 }
             }
-            scopeProperty == Scope.Major() -> "${lastSemVer.inc(Increase.Major, incStage)}"
-            scopeProperty == Scope.Minor() -> "${lastSemVer.inc(Increase.Minor, incStage)}"
-            scopeProperty == Scope.Patch() -> "${lastSemVer.inc(Increase.Patch, incStage)}"
+            scopeProperty == Scope.Major() -> "${lastSemver.inc(Increase.Major, incStage)}"
+            scopeProperty == Scope.Minor() -> "${lastSemver.inc(Increase.Minor, incStage)}"
+            scopeProperty == Scope.Patch() -> "${lastSemver.inc(Increase.Patch, incStage)}"
             scopeProperty == Scope.Auto() -> {
                 when {
-                    !isThereVersionTag(tagPrefix) -> "$lastSemVer"
-                    incStage.isEmpty() -> "${lastSemVer.inc(Increase.Patch, incStage)}"
-                    else -> "${lastSemVer.inc(stageName = incStage)}"
+                    !isThereVersionTag(tagPrefix) -> "$lastSemver"
+                    incStage.isEmpty() -> "${lastSemver.inc(Increase.Patch, incStage)}"
+                    else -> "${lastSemver.inc(stageName = incStage)}"
                 }
             }
-            isCreatingSemverTag && !isThereVersionTag(tagPrefix) -> "$lastSemVer"
-            else -> "${lastSemVer.inc(stageName = incStage)}"
+            isCreatingSemverTag && !isThereVersionTag(tagPrefix) -> "$lastSemver"
+            else -> "${lastSemver.inc(stageName = incStage)}"
         }.also {
             if (it.contains("final", true) || it.contains("auto", true)) {
                 error("`stage` plus `scope` combination is broken, please report it")
@@ -77,18 +81,17 @@ internal fun Git.calculatedVersion(
 
 internal fun Git.calculateAdditionalVersionData(
     tagPrefix: String,
-    mockDate: Date?,
     checkClean: Boolean = true,
 ): String {
     val isClean = status().call().isClean || !checkClean
 
-    val commitsBetweenCurrentAndLastTagCommit =
+    val commitsBetweenCurrentAndLastTagCommit: List<GitRef.Commit> =
         commitsBetweenTwoCommitsIncludingLastExcludingFirst(
             lastCommitInCurrentBranch,
             lastVersionCommitInCurrentBranch(tagPrefix)
         )
 
-    val noVersionTags = isThereVersionTag(tagPrefix).not()
+    val noVersionTags: Boolean = isThereVersionTag(tagPrefix).not()
 
     val additionalData: String =
         commitsBetweenCurrentAndLastTagCommit.run {
@@ -97,15 +100,17 @@ internal fun Git.calculateAdditionalVersionData(
                     ".$size+${headCommit.commit.hash.take(DEFAULT_SHORT_HASH_LENGTH)}"
                 }
                 noVersionTags && !isClean -> {
-                    ".$size+${timestamp(mockDate)}"
-                }
-                isClean && isEmpty() -> {
-                    ""
+                    ".$size+${headCommit.commit.hash.take(DEFAULT_SHORT_HASH_LENGTH)}+DIRTY"
                 }
                 isNotEmpty() && isClean -> {
                     ".$size+${first().hash.take(DEFAULT_SHORT_HASH_LENGTH)}"
                 }
-                else -> ".$size+${timestamp(mockDate)}"
+                isEmpty() && isClean -> {
+                    ""
+                }
+                else -> {
+                    ".$size+DIRTY"
+                }
             }
         }