+++ CHANGELOG.md
@@ -4,11 +4,28 @@
 
 ### Added
 
+- `semverPrint` task which prints the version and generates the `build/semver/version.txt`.
+- `semver` plugin extension which has `tagPrefix` to indicate the project tag prefix.
+- configuration cache support
+- project isolation support
+
 ### Changed
+
+- the plugin must be applied individually to each project instead of only in the root project.
+- `project.version` is now a `LazyVersion`, to get the string use `toString` method.
+- `createSemverTag` name to `semverCreateTag`.
+- `pushSemverTag` name to `semverPushTag`.
+- `semver.tagPrefix` is no longer used to indicate a project version. Now it is used to filter which
+  projects are going to bump the version based on it and the value in the property `tagPrefix` in
+  the semver plugin extension.
 
 ### Deprecated
 
 ### Removed
+
+- applying the plugin to the root project only configures the root project and not all projects.
+- `build/semver/version.txt` is no longer generated in configuration phase.
+- the old way to change the version in multi-project builds (`-PprojectName:semver.scope=patch"`).
 
 ### Fixed